% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_iaml.R
\name{mlr_measures_iaml_selected_non_monotone}
\alias{mlr_measures_iaml_selected_non_monotone}
\alias{MeasureIAMLSelectedNonMonotone}
\title{IAML Selected Non Monotone Features Measure}
\description{
Measures the number of non monotone features by extracting it from an appropriate \link[mlr3:Learner]{mlr3::Learner}
If parameter \code{normalize} is set to \code{TRUE}, the relative number of non monotone features instead of the absolute
number of non monotone is returned.
Note that the models must be stored to be able to extract this information.
This measure is typically only used in combination with \link{TunerIAML} because it requires parameters of the learner to
have additional attributes set and will not work in other workflows.

This measure requires the \link[mlr3:Task]{mlr3::Task} and the \link[mlr3:Learner]{mlr3::Learner} for scoring.
}
\section{Dictionary}{

This \link{Measure} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary} \link{mlr_measures} or with the associated sugar function \code{\link[=msr]{msr()}}:\preformatted{mlr_measures$get("iaml_selected_non_monotone")
msr("iaml_selected_non_monotone")
}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{NA}
\item Range: \eqn{[0, \infty)}{[0, Inf)}
\item Minimize: TRUE
\item Average: macro
\item Required Prediction: \dQuote{response}
\item Required Packages: \CRANpkg{mlr3}
}
}

\section{Parameters}{
\tabular{llll}{
   Id \tab Type \tab Default \tab Levels \cr
   normalize \tab logical \tab FALSE \tab TRUE, FALSE \cr
   monotone_id \tab untyped \tab - \tab  \cr
}


\describe{
\item{\code{normalize}}{\code{logical(1)}\cr
Should the relative number of non monotone features instad of the absolute number be returned?
}
\item{\code{monotone_id}}{\code{character(1)}\cr
ID of param in Learner that sets monotonicity constraints.}
}
}

\section{Super class}{
\code{\link[mlr3:Measure]{mlr3::Measure}} -> \code{MeasureIAMLSelectedNonMonotone}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MeasureIAMLSelectedNonMonotone$new()}}
\item \href{#method-clone}{\code{MeasureIAMLSelectedNonMonotone$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="aggregate">}\href{../../mlr3/html/Measure.html#method-aggregate}{\code{mlr3::Measure$aggregate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="format">}\href{../../mlr3/html/Measure.html#method-format}{\code{mlr3::Measure$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="help">}\href{../../mlr3/html/Measure.html#method-help}{\code{mlr3::Measure$help()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="print">}\href{../../mlr3/html/Measure.html#method-print}{\code{mlr3::Measure$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="score">}\href{../../mlr3/html/Measure.html#method-score}{\code{mlr3::Measure$score()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureIAMLSelectedNonMonotone$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureIAMLSelectedNonMonotone$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
